requirements.txt
openai==1.82.0
tiktoken==0.9.0
langchain==0.3.25
faiss-cpu==1.11.0
python-dotenv==1.1.0
beautifulsoup4==4.13.4
langchain-openai==0.3.18
langchain-community==0.3.24
